# api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Reign Design Api",
    "description": "Api service for Reign Design",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "anibalrodriguez1990@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "tags": [
    {
      "name": "posts",
      "description": "Posts endpoints"
    }
  ],
  "paths": {
    "/api/posts": {
      "get": {
        "tags": ["posts"],
        "summary": "Endpoint to get all posts",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostResponse"
                  }
                }
              }
            }
          },
          "403": {
            "description": "param error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "path api not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/refresh": {
      "post": {
        "tags": ["posts"],
        "summary": "Endpoint to manage posts",
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "403": {
            "description": "param error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "path api not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{id}": {
      "delete": {
        "tags": ["posts"],
        "summary": "Endpoint to delete posts",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Id from post"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {}
            }
          },
          "403": {
            "description": "param error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "path api not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PostResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "objectID": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          },
          "created_at": {
            "type": "number"
          },
          "author": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "__v": {
            "type": "number"
          }
        },
        "xml": {
          "name": "PostResponse"
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "xml": {
          "name": "ErrorResponse"
        }
      }
    }
  }
}